rule shell python script
